+++ astropy/io/fits/tests/test_core.py
@@ -7,6 +7,7 @@ import zipfile
 import numpy as np
 
 from astropy.io import fits
+from astropy.io.fits.convenience import _getext
 from astropy.tests.helper import pytest, raises
 
 from . import FitsTestCase
@@ -192,14 +193,83 @@ class TestCore(FitsTestCase):
         hdu = fits.ImageHDU()
         hdu.verify('foobarbaz')
 
+    def test_getext(self):
+        """
+        Test the various different ways of specifying an extension header in
+        the convenience functions.
+        """
+
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', 1)
+        assert_equal(ext, 1)
+        assert_raises(ValueError, _getext, self.data('test0.fits'), 'readonly',
+                      1, 2)
+        assert_raises(ValueError, _getext, self.data('test0.fits'), 'readonly',
+                      (1, 2))
+        assert_raises(ValueError, _getext, self.data('test0.fits'), 'readonly',
+                      'sci', 'sci')
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      1, 2, 3)
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', ext=1)
+        assert_equal(ext, 1)
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', ext=('sci', 2))
+        assert_equal(ext, ('sci', 2))
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      1, ext=('sci', 2), extver=3)
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      ext=('sci', 2), extver=3)
+
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', 'sci')
+        assert_equal(ext, ('sci', 0))
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', 'sci', 1)
+        assert_equal(ext, ('sci', 1))
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', ('sci', 1))
+        assert_equal(ext, ('sci', 1))
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', 'sci',
+                          extver=1, do_not_scale_image_data=True)
+        assert_equal(ext, ('sci', 1))
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      'sci', ext=1)
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      'sci', 1, extver=2)
+
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', extname='sci')
+        assert_equal(ext, ('sci', 0))
+        hl, ext = _getext(self.data('test0.fits'), 'readonly', extname='sci',
+                          extver=1)
+        assert_equal(ext, ('sci', 1))
+        assert_raises(TypeError, _getext, self.data('test0.fits'), 'readonly',
+                      extver=1)
+
+
+class TestConvenienceFunctions(PyfitsTestCase):
+    def test_writeto(self):
+        """
+        Simple test for writing a trivial header and some data to a file
+        with the `writeto()` convenience function.
+        """
+
+        data = np.zeros((100,100))
+        header = pyfits.Header()
+        pyfits.writeto(self.temp('array.fits'), data, header=header,
+                       clobber=True)
+        hdul = pyfits.open(self.temp('array.fits'))
+        assert_equal(len(hdul), 1)
+        assert_true((data == hdul[0].data).all())
+
 
 class TestFileFunctions(FitsTestCase):
-    """Tests various basic I/O operations, specifically in the
+    """
+    Tests various basic I/O operations, specifically in the
     astropy.io.fits.file._File class.
     """
 
     def test_open_gzipped(self):
         assert len(fits.open(self._make_gzip_file())) == 5
+
+    def test_detect_gzipped(self):
+        """Test detection of a gzip file when the extension is not .gz."""
+
+        assert_equal(len(pyfits.open(self._make_gzip_file('test0.fz'))), 5)
 
     def test_open_gzipped_writeable(self):
         """Opening gzipped files in a writeable mode should fail."""
@@ -211,6 +281,12 @@ class TestFileFunctions(FitsTestCase):
     def test_open_zipped(self):
         assert len(fits.open(self._make_zip_file())) == 5
 
+    def test_detect_zipped(self):
+        """Test detection of a zip file when the extension is not .zip."""
+
+        zf = self._make_zip_file(filename='test0.fz')
+        assert_equal(len(pyfits.open(zf)), 5)
+
     def test_open_zipped_writeable(self):
         """Opening zipped files in a writeable mode should fail."""
 
@@ -220,7 +296,8 @@ class TestFileFunctions(FitsTestCase):
 
     @raises(IOError)
     def test_open_multipe_member_zipfile(self):
-        """Opening zip files containing more than one member files should fail
+        """
+        Opening zip files containing more than one member files should fail
         as there's no obvious way to specify which file is the FITS file to
         read.
         """
@@ -283,8 +360,8 @@ class TestFileFunctions(FitsTestCase):
 
         assert old_mode == os.stat(filename).st_mode
 
-    def _make_gzip_file(self):
-        gzfile = self.temp('test0.fits.gz')
+    def _make_gzip_file(self, filename='test0.fits.gz'):
+        gzfile = self.temp(filename)
         with open(self.data('test0.fits'), 'rb') as f:
             gz = gzip.open(gzfile, 'wb')
             gz.write(f.read())
@@ -292,8 +369,8 @@ class TestFileFunctions(FitsTestCase):
 
         return gzfile
 
-    def _make_zip_file(self, mode='copyonwrite'):
-        zfile = zipfile.ZipFile(self.temp('test0.fits.zip'), 'w')
+    def _make_zip_file(self, mode='copyonwrite', filename='test0.fits.zip'):
+        zfile = zipfile.ZipFile(self.temp(filename), 'w')
         zfile.write(self.data('test0.fits'))
         zfile.close()
 
@@ -310,8 +387,9 @@ class TestStreamingFunctions(FitsTestCase):
 
     @raises(ValueError)
     def test_streaming_hdu_file_wrong_mode(self):
-        """Test that streaming an HDU to a file opened in the wrong mode
-        fails as expected.
+        """
+        Test that streaming an HDU to a file opened in the wrong mode fails as
+        expected.
         """
 
         with open(self.temp('new.fits'), 'wb') as f:
@@ -356,21 +434,6 @@ class TestStreamingFunctions(FitsTestCase):
         with open(self.temp('new.fits'), 'ab+') as f:
             shdu = self._make_streaming_hdu(f)
             shdu.write(arr)
-
-    # TODO: This test is temporarily borrowed from the header-refactoring
-    # branch as a regression test for ticket #69; it can be removed when the
-    # header-refactoring branch is merged into trunk
-    def test_update_comment(self):
-        hdul = fits.open(self.data('arange.fits'))
-        hdul[0].header.update('FOO', 'BAR', 'BAZ')
-        hdul.writeto(self.temp('test.fits'))
-
-        hdul = fits.open(self.temp('test.fits'), mode='update')
-        hdul[0].header.ascard['FOO'].comment = 'QUX'
-        hdul.close()
-
-        hdul = fits.open(self.temp('test.fits'))
-        assert hdul[0].header.ascard['FOO'].comment == 'QUX'
 
     def _make_streaming_hdu(self, fileobj):
         hd = fits.Header()